NEWS-1 Implement tabs to display list of years and grab proper issues based on year

+++ amf-registration/modules/newsletter/newsletter-service/src/main/java/com/liferay/amf/newsletter/service/impl/IssueLocalServiceImpl.java
@@ -106,7 +106,7 @@ public class IssueLocalServiceImpl extends IssueLocalServiceBaseImpl {
 
 		DynamicQuery issueQuery = DynamicQueryFactoryUtil.forClass(Issue.class, classLoader)
 				.setProjection(ProjectionFactoryUtil.sqlGroupProjection("year(issueDate) as year","year",
-						new String[] { "year" }, new Type[] { Type.LONG }));
+						new String[] { "year" }, new Type[] { Type.INTEGER }));
 		List<Integer> issueYears = issueLocalService.dynamicQuery(issueQuery);
 		return issueYears;
 	}
